svm: add shared.util.NumUtil
