+++ substratevm/src/com.oracle.svm.core.posix/src/com/oracle/svm/core/posix/linux/LinuxImageHeapProvider.java
@@ -81,15 +81,16 @@ import com.oracle.svm.core.posix.headers.Fcntl;
 import com.oracle.svm.core.posix.headers.Unistd;
 import com.oracle.svm.core.util.PointerUtils;
 import com.oracle.svm.core.util.UnsignedUtils;
-import com.oracle.svm.shared.Uninterruptible;
 import com.oracle.svm.guest.staging.c.CGlobalData;
 import com.oracle.svm.guest.staging.c.CGlobalDataFactory;
 import com.oracle.svm.hosted.imagelayer.ImageLayerSectionFeature;
 import com.oracle.svm.hosted.imagelayer.LayeredDispatchTableFeature;
+import com.oracle.svm.shared.Uninterruptible;
 import com.oracle.svm.shared.singletons.traits.BuiltinTraits.AllAccess;
 import com.oracle.svm.shared.singletons.traits.BuiltinTraits.SingleLayer;
 import com.oracle.svm.shared.singletons.traits.SingletonLayeredInstallationKind.InitialLayerOnly;
 import com.oracle.svm.shared.singletons.traits.SingletonTraits;
+import com.oracle.svm.shared.util.NumUtil;
 import com.oracle.svm.shared.util.VMError;
 
 import jdk.graal.compiler.nodes.NamedLocationIdentity;
@@ -257,7 +258,7 @@ public class LinuxImageHeapProvider extends AbstractImageHeapProvider {
 
         int offset = startOffset;
         long bitmapWordCountAsLong = data.readLong(offset);
-        int bitmapWordCount = UninterruptibleUtils.NumUtil.safeToInt(bitmapWordCountAsLong);
+        int bitmapWordCount = NumUtil.safeToInt(bitmapWordCountAsLong);
         offset += Long.BYTES;
         if (addend.equal(0)) {
             /* Nothing to do. */
@@ -298,7 +299,7 @@ public class LinuxImageHeapProvider extends AbstractImageHeapProvider {
         int referenceSize = ConfigurationValues.getObjectLayout().getReferenceSize();
         int offset = startOffset;
         long countAsLong = patches.readLong(offset);
-        int count = UninterruptibleUtils.NumUtil.safeToInt(countAsLong);
+        int count = NumUtil.safeToInt(countAsLong);
         offset += Long.BYTES;
         int endOffset = offset + count * Integer.BYTES;
         while (offset < endOffset) {
